Implement LongClickDeleteListener
Nicer than having to reimplement the same thing in every activity.
